requirements.txt
aiohttp==3.9.1
websockets==12.0
python-telegram-bot==21.4
fake-useragent==1.5.1
python-dotenv==1.0.0